requirements.txt
fastapi==0.121.1
uvicorn[standard]==0.38.0
google-generativeai==0.8.5
python-dotenv==1.0.1
pydantic==2.12.4
requests==2.32.5

